refactor: table buttons
